requirements.txt
requests
pandas
google-cloud-bigquery
functions-framework
dbt-bigquery
flask
